+++ ChangeLog.md
@@ -11,7 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 
-- added function merge, assoc-in, get-in
+- added function keep, merge, assoc-in, get-in
 - added function update, vec, difference
 - added function every? and any? operating on sequential collections
 - added associative destructuring